CRUISE: Fix for ScriptInstance code changes to prevent invalid accesses.
The previous commit fixed most of the issues, but there were a few
missed. The ScriptInstance was originally zeroed at allocation, so
the _type field needed to be assigned zero in the new constructor.
To do this cleanly, this value (which was previously implicit) was
added to the enumeration. Also, the correction to the removeAllScripts
iterator usage was slightly incorrect with iter being accessed after
erasure and the change to avoid the pointer operator being unecessary.