CRUISE: Fix double free in _ptr when saving BackgroundInCrust objects.
BackgroundInCrust now has a destructor which frees the _ptr member.
Thus by creating a temporary BackgroundInCrust object while saving and
assigning it from a object in the current list, this results in the
_ptr member of the last list object being freed as a side effect, when
the function ends and local BackgroundInCrust object is destructed.

To avoid this, cleared the _ptr of the local temporary object to NULL
at the end of the function.