OPENGL: Modify xBRZ Header Paths and Add to Make Rules.
This fails to compile as the ppl.h system header is for the Microsoft
Parallel Patterns Library which is not a "standard" header and thus not
present on Linux.

Apart from this, the code for both the scaler and patches generate a
large number of errors and warnings currently as well. Suspect that this
code will only compile for Windows with MSVC or possibly Min-GW.