IMMORTAL: Fix Various GCC Compiler Warnings
Some changes to this code are not functional, but sufficient to fix
the warnings i.e. adding of fallthrough comments, removal of extraneous
semi-colons. However, some of these changes will require analysis and
playtesting to check that these are correct i.e. fixes to the size of
structure variables, additional of breaks in some switch statements.